backend/requirements.txt
Flask==3.1.0
python-dotenv==0.21.0
Flask-Cors==3.0.10
